ProfileSetup moved to screens and renamed ProfileSetupScreens in the course of refactoring the files to align with phase 1 of the project
